Windowed LRU updates
Optimize the "get" operation by keeping track of LRU entry promotion and
moving to LRU list front only after 1 second has elapsed. The rationale
here is that recently used entries will anyway be at the top of the
list. With windowed updates, the "get" operation only has to take the
read lock (which is more scalable) for majority of the operations.

Before:

  [penberg@nero golru]$ go test -bench=.
  goos: linux
  goarch: amd64
  pkg: github.com/penberg/golru
  BenchmarkLRU-8           5000000               311 ns/op
  PASS
  ok      github.com/penberg/golru        1.873s

After:

  [penberg@nero golru]$ go test -bench=.
  goos: linux
  goarch: amd64
  pkg: github.com/penberg/golru
  BenchmarkLRU-8          30000000                44.1 ns/op
  PASS
  ok      github.com/penberg/golru        1.369s
  [penberg@nero golru]$ git revert HEAD